+++ website/about/changelog.md
@@ -10,6 +10,7 @@ icon: material/file-document-edit
 !!! quote ""
     - Move website to tremeschin.com domain
     - Fix `Scene.resolution = (w, h)` using positionals in `Scene.resize`
+    - Simplify remove `scene._scale` field in favor of `.main()` or `.resize()` argument
 
 ### 📦 v0.10.0 <small>March 12, 2026</small> {#v0.10.0}